+++ tests/opentelemetry/test_tracing_sync.py
@@ -314,7 +314,7 @@ class TestNoSpansWhenDisabled:
     def test_no_spans_without_enable_tracing(self):
         """Without enable_tracing(), the registry uses noop — no spans are created."""
 
-        from tests.tracing.conftest import _exporter
+        from tests.opentelemetry.conftest import _exporter
 
         _registry.set_create_span(None)
         _registry.set_metadata_hook(None)
